Add "Context bindings - Resolving bindings"
